chore: setup component folder structure
