+++ src/stories/buttons/button/index.tsx
@@ -1,6 +1,6 @@
 import { Button as ZendeskButton } from "@zendeskgarden/react-buttons";
 import styled from "styled-components";
-import { paddingDefault } from "./_shared";
+import { paddingDefault } from "../shared/_shared";
 import { ButtonArgs } from "./_types";
 
 const UgButton = styled(ZendeskButton)`
