+++ src/stories/buttons/shared/_shared.tsx
@@ -1,0 +1,7 @@
+import { css } from "styled-components";
+
+const paddingDefault = css`
+  padding: 0.714rem 1.071rem;
+`;
+
+export { paddingDefault };
